backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
edge-tts>=6.1.0
google-genai>=1.0.0
openai>=1.50.0
